requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pymongo==4.5.0
twilio==8.5.0
gunicorn==21.2.0
openai==1.70.0
gTTS==2.3.2
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
blinker==1.6.2
click==8.1.7
selenium==4.15.0
webdriver-manager==4.0.1
pyotp==2.9.0
bson==0.5.10